update max population
arg as null, defaults to the sum of 2 times the offspring counts
plus the length of the species set times two

Populate new generation is updated for keep all

two pairs selected instead of one.  Prevents cancelling of
prior progress if one of the two new individuals is selected as
a parent.  Can be assumed to happen once at times but
rare that it happens twice, less so since the next generation
will have two times the species offspring count instead of just
one times the offspring count